HGW-Flask/requirements.txt
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==2.0.3
python-dotenv==1.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.41
click==8.1.8
itsdangerous==2.2.0
colorama==0.4.6
typing_extensions==4.14.0
blinker==1.9.0
